Create Freq_Hop_Entropv2.ino
Here's a refactored version of the previous code that includes the necessary parts of the Entropy library directly in the sketch. This code demonstrates a simple form of a pseudo-random frequency hopping system. Remember to adhere to local laws and regulations when it comes to radio frequency transmissions.

This code includes a simplified version of the Entropy library as the EntropyClass, which provides basic random number generation using the watchdog timer. The refactored code initializes the EntropyClass object in the setup function, and then continually hops frequencies every second in the loop function. The getRandomFrequency function generates a random frequency within the defined range using the EntropyClass object.

The actual implementation of the setRadioFrequency function will depend on the specific radio being used. This function should control the radio's frequency based on the input argument.

Please note that this code is a simple example and does not include error checking, exception handling, or other robustness measures you might need in a real-world application.

see also: https://github.com/jypma/AvrLib