Create Simple_randnum.cpp
a simpler library for generating random numbers on an Arduino Uno. It's worth noting that the Arduino Uno already has a built-in random number generator, but it's not truly random. It's a pseudo-random number generator, meaning it generates numbers that appear random but are actually deterministic.

The randomSeed(analogRead(A0)) function is often used to seed the random number generator with noise from an unconnected analog pin to make the random numbers less predictable.

Here's a simple library that uses this technique to generate 8, 16, and 32-bit random numbers:

This is a lot simpler than the previous library, but it's worth noting that the random numbers it generates are less random, particularly if the analog pin doesn't pick up much noise. If you need high-quality random numbers for cryptography or simulations, you may want to use a more sophisticated method like the one in the previous library.